Revert "ion: change ion_user_handle_t typedef from int to long, to support 64/32 bit app."
This reverts commit 2760b0f3db1059c2a852c550d85347c4c101e882.

+++ drivers/staging/android/uapi/ion.h
@@ -20,7 +20,7 @@
 #include <linux/ioctl.h>
 #include <linux/types.h>
 
-typedef long ion_user_handle_t;
+typedef int ion_user_handle_t;
 
 /**
  * enum ion_heap_types - list of all possible types of heaps
